Move KeyFormat trait to upper level and clean up docstrings.
